+++ UmbrellaPipeline/path_generation/grid.py
@@ -119,8 +119,7 @@ class Grid:
             psf = psf
 
         inx = get_residue_indices(psf.atom_list)
-        if psf.boxVectors == None:
-            min_c = gen_pbc_box(psf, pdb)
+        min_c = gen_pbc_box(psf, pdb)
 
         n = [
             round(psf.boxLengths[0] / gridsize),
